[python] Consolidate `_set_coords` into single function
