+++ apis/python/src/tiledbsoma/_util.py
@@ -9,14 +9,27 @@ import pathlib
 import time
 import urllib.parse
 from itertools import zip_longest
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Type, TypeVar, Union, cast
+from typing import (
+    Any,
+    Dict,
+    List,
+    Mapping,
+    Optional,
+    Sequence,
+    Tuple,
+    Type,
+    TypeVar,
+    Union,
+    cast,
+)
 
+import numpy as np
 import pyarrow as pa
 import somacore
 from somacore import options
 
 from . import pytiledbsoma as clib
-from ._types import OpenTimestamp, Slice, is_slice_of
+from ._types import OpenTimestamp, Slice, is_nonstringy_sequence, is_slice_of
 from .options._tiledb_create_write_options import (
     TileDBCreateOptions,
     _ColumnConfig,
@@ -428,3 +441,169 @@ def _build_filter_list(
                     )
         filter_list.append(filter)
     return json.dumps(filter_list) if return_json else filter_list
+
+
+def _cast_domainish(domainish: List[Any]) -> Tuple[Tuple[object, object], ...]:
+    result = []
+    for slot in domainish:
+
+        arrow_type = slot[0].type
+        if pa.types.is_timestamp(arrow_type):
+            pandas_type = np.dtype(arrow_type.to_pandas_dtype())
+            result.append(
+                tuple(
+                    pandas_type.type(e.cast(pa.int64()).as_py(), arrow_type.unit)
+                    for e in slot
+                )
+            )
+        else:
+            result.append(tuple(e.as_py() for e in slot))
+
+    return tuple(result)
+
+
+def _set_coords(sr: clib.SOMAArray, coords: options.SparseNDCoords) -> None:
+    if not is_nonstringy_sequence(coords):
+        raise TypeError(
+            f"coords type {type(coords)} must be a regular sequence,"
+            " not str or bytes"
+        )
+
+    if len(coords) > len(sr.dimension_names):
+        raise ValueError(
+            f"coords ({len(coords)} elements) must be shorter than ndim"
+            f" ({len(sr.dimension_names)})"
+        )
+
+    for i, coord in enumerate(coords):
+        _set_coord(i, sr, coord)
+
+
+def _set_coord(dim_idx: int, sr: clib.SOMAArray, coord: object) -> None:
+    if coord is None:
+        return
+
+    dim = sr.schema.field(dim_idx)
+    dom = _cast_domainish(sr.domain())[dim_idx]
+
+    if isinstance(coord, (str, bytes)):
+        sr.set_dim_points_string_or_bytes(dim.name, [coord])
+        return
+
+    if isinstance(coord, (pa.Array, pa.ChunkedArray)):
+        sr.set_dim_points_arrow(dim.name, coord)
+        return
+
+    if isinstance(coord, (Sequence, np.ndarray)):
+        _set_coord_by_py_seq_or_np_array(sr, dim, coord)
+        return
+
+    if isinstance(coord, int):
+        sr.set_dim_points_int64(dim.name, [coord])
+        return
+
+    # Note: slice(None, None) matches the is_slice_of part, unless we also check
+    # the dim-type part
+    if (
+        is_slice_of(coord, str) or is_slice_of(coord, bytes)
+    ) and pa_types_is_string_or_bytes(dim.type):
+        validate_slice(coord)
+        dim_type = type(dom[0])
+        # A ``None`` or empty start is always equivalent to empty str/bytes.
+        start = coord.start or dim_type()
+        if coord.stop is None:
+            # There's no way to specify "to infinity" for strings.
+            # We have to get the nonempty domain and use that as the end.\
+            ned = _cast_domainish(sr.non_empty_domain())
+            _, stop = ned[dim_idx]
+        else:
+            stop = coord.stop
+        sr.set_dim_ranges_string_or_bytes(dim.name, [(start, stop)])
+        return
+
+    # Note: slice(None, None) matches the is_slice_of part, unless we also check
+    # the dim-type part.
+    if is_slice_of(coord, np.datetime64) and pa.types.is_timestamp(dim.type):
+        validate_slice(coord)
+
+        # These timestamp types are stored in Arrow as well as TileDB as 64-bit
+        # integers (with distinguishing metadata of course). For purposes of the
+        # query logic they're just int64.
+        ts_dom = pa.array(dom, type=dim.type).cast(pa.int64())
+
+        if coord.start is not None:
+            istart = coord.start.astype("int64")
+        else:
+            istart = ts_dom[0].as_py()
+
+        if coord.stop is not None:
+            istop = coord.stop.astype("int64")
+        else:
+            istop = ts_dom[1].as_py()
+
+        sr.set_dim_ranges_int64(dim.name, [(istart, istop)])
+        return
+
+    if isinstance(coord, slice):
+        validate_slice(coord)
+        if coord.start is None and coord.stop is None:
+            return
+        _set_coord_by_numeric_slice(sr, dim, dom, coord)
+        return
+
+    raise TypeError(f"unhandled type {dim.type} for index column named {dim.name}")
+
+
+def _set_coord_by_py_seq_or_np_array(
+    sr: clib.SOMAArray, dim: pa.Field, coord: object
+) -> None:
+    if isinstance(coord, np.ndarray):
+        if coord.ndim != 1:
+            raise ValueError(
+                f"only 1D numpy arrays may be used to index; got {coord.ndim}"
+            )
+
+    try:
+        set_dim_points = getattr(sr, f"set_dim_points_{dim.type}")
+    except AttributeError:
+        # We have to handle this type specially below
+        pass
+    else:
+        set_dim_points(dim.name, coord)
+        return
+
+    if pa_types_is_string_or_bytes(dim.type):
+        sr.set_dim_points_string_or_bytes(dim.name, coord)
+        return
+
+    if pa.types.is_timestamp(dim.type):
+        if not isinstance(coord, (tuple, list, np.ndarray)):
+            raise ValueError(
+                f"unhandled coord type {type(coord)} for index column named {dim.name}"
+            )
+        icoord = [
+            int(e.astype("int64")) if isinstance(e, np.datetime64) else e for e in coord
+        ]
+        sr.set_dim_points_int64(dim.name, icoord)
+        return
+
+    raise ValueError(f"unhandled type {dim.type} for index column named {dim.name}")
+
+
+def _set_coord_by_numeric_slice(
+    sr: clib.SOMAArray, dim: pa.Field, dom: Tuple[object, object], coord: Slice[Any]
+) -> None:
+    try:
+        lo_hi = slice_to_numeric_range(coord, dom)
+    except NonNumericDimensionError:
+        return
+
+    if not lo_hi:
+        return
+
+    try:
+        set_dim_range = getattr(sr, f"set_dim_ranges_{dim.type}")
+        set_dim_range(dim.name, [lo_hi])
+        return
+    except AttributeError:
+        return
